requirements.txt
###### GraphQL ######
graphene-django
